requirements.txt
aiohttp
pillow
pyyaml
requests
qrcode